feat: 资源设置修改value字段类型

+++ HymsonMES/src/Libraries/Hymson.MES.Services/Dtos/Process/ProcResourceConfigResDto.cs
@@ -87,7 +87,7 @@ namespace Hymson.MES.Services.Dtos.Process
        /// <summary>
         /// 设置值
         /// </summary>
-        public int? Value { get; set; }
+        public long? Value { get; set; }
 
         /// <summary>
         /// Id
@@ -108,7 +108,7 @@ namespace Hymson.MES.Services.Dtos.Process
         /// <summary>
         /// 设置值
         /// </summary>
-        public int? Value { get; set; }
+        public long? Value { get; set; }
 
         /// <summary>
         /// Id
